+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-## v0.9.1 - UNRELEASED
+## v0.9.2 - Unreleased
+
+* Add lenient support to Zone.add_record, allows populate from providers that
+  have allowed/created invalid data and situations where a sub-zone is being
+  extracted from a parent, but the records still exist in the remote provider.
+
+## v0.9.1 - 2018-05-21 - Going backwards with setup.py
 
 ### NOTICE
 
@@ -11,6 +17,10 @@ all health checks are passing before the first sync with `--doit`. See
 
 * Major update to geo healthchecks to allow configuring host (header), path,
   protocol, and port [#67](https://github.com/github/octodns/pull/67)
+* SSHFP algorithm type 4
+* NS1 and DNSimple support skipping unsupported record types
+* Revert back to old style setup.py &amp; requirements.txt, setup.cfg was
+  causing too much pita
 
 ## v0.9.0 - 2018-03-26 - Way too long since we last met